[fix]update version of springboot to 2.1.4 from 2.0.4 and fix SampleRepository

+++ eternal-app/src/main/java/com/eternal/web/repository/SampleTrnRepository.java
@@ -1,9 +1,10 @@
 package com.eternal.web.repository;
 
+import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.eternal.web.entity.SampleTrnEntity;
 @Repository
-public interface SampleTrnRepository extends JpaRepository<SampleTrnEntity, Long> {
+public interface SampleTrnRepository extends JpaRepository<SampleTrnEntity, BigDecimal> {
 
 }
